requirements.txt
kafka-python
dash
plotly
yfinance
tqdm